Split hitTest.ts (1072 lines) into 5 focused modules
- hitTestUtils.ts: shared primitives (HitResult, isSelectable, geometry)
- hitTestPoint.ts: single-point hit testing
- hitTestArea.ts: box, lasso, outside-box, line selection
- hitTestConnectivity.ts: connection select with wireByUuid Map
- hitTestFloodFill.ts: spatial-indexed wire flood fill (private)
- hitTest.ts: barrel re-export (no callers changed)
